fixup! Fix pointed codes
